ISSUE-18: Added test cases to validate validation logic of paritioning configuration, data_type and field parameters
